backend/pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "disclosure-backend"
version = "0.1.0"
description = "Backend service for the disclosure comparison tool"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.28",
    "celery>=5.3",
    "redis>=5.0",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0",
    "PyMuPDF>=1.23",
    "pdfplumber>=0.10",
    "pdf2image>=1.17",
    "Pillow>=10.0",
    "sentence-transformers>=2.6",
    "openai>=1.30",
    "httpx>=0.27",
    "structlog>=24.1",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["storage*", "templates*", "tests*"]

[tool.uvicorn]
factory = true
app = "app.main:create_app"
reload = true
host = "0.0.0.0"
port = 8000
